+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.3.8 (unreleased)
 
+- Colorize output
 - Fixed error when unable to parse view definitions
 
 ## 0.3.7 (2020-07-10)